+++ data/src/test/java/com/clloret/days/data/repository/AppEventRepositoryTest.java
@@ -25,6 +25,7 @@ import io.reactivex.Observable;
 import io.reactivex.observers.TestObserver;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 import okhttp3.mockwebserver.MockWebServer;
 import org.joda.time.DateTime;
 import org.junit.After;
@@ -44,7 +45,10 @@ public class AppEventRepositoryTest {
   private static final String API_KEY = "api_key";
   private static final String BASE = "base";
   private static final Correspondence<Event, String> EVENT_HAS_NAME =
-      Correspondence.from((actual, expected) -> actual.getName().equals(expected), "contains");
+      Correspondence.from((actual, expected) -> actual != null && actual.getName().equals(expected),
+          "contains");
+  @Rule
+  public final RxImmediateSchedulerRule schedulers = new RxImmediateSchedulerRule();
   private final MockWebServer server = new MockWebServer();
   private final MockUtils mockUtils = new MockUtils(server);
   private final DbEventDataMapper dbEventDataMapper = new DbEventDataMapper();
@@ -52,9 +56,6 @@ public class AppEventRepositoryTest {
   private EventRepository eventRepository;
   private RoomEventRepository roomEventRepository;
   private DaysDatabase db;
-
-  @Rule
-  public final RxImmediateSchedulerRule schedulers = new RxImmediateSchedulerRule();
 
   @Before
   public void setUp() throws Exception {
@@ -121,7 +122,7 @@ public class AppEventRepositoryTest {
         .assertComplete()
         .assertNoErrors()
         .assertValueCount(1)
-        .assertValue(event);
+        .assertValue(result -> Objects.equals(result.getId(), event.getId()));
   }
 
   @Test