+++ domain/src/main/java/com/clloret/days/domain/entities/Identifiable.java
@@ -1,8 +1,0 @@
-package com.clloret.days.domain.entities;
-
-public interface Identifiable {
-
-  String getId();
-
-  void setId(String id);
-}