+++ domain/src/main/java/com/clloret/days/domain/events/order/EventSortable.kt
@@ -1,0 +1,9 @@
+package com.clloret.days.domain.events.order
+
+import java.util.*
+
+interface EventSortable {
+  val isFavorite: Boolean
+  val name: String?
+  val date: Date?
+}